Stream files splitted into separated files
